Instead of using an engine from the SSLHelper, get an SslContext and let the channel configurator create an SslHandler from the SslContext
