+++ src/main/java/io/vertx/core/net/impl/NetSocketImpl.java
@@ -12,9 +12,11 @@
 package io.vertx.core.net.impl;
 
 import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufAllocator;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.*;
 import io.netty.handler.ssl.SniHandler;
+import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.util.CharsetUtil;
 import io.netty.util.ReferenceCounted;
@@ -353,14 +355,12 @@ public class NetSocketImpl extends ConnectionBase implements NetSocketInternal {
         }
       });
       if (remoteAddress != null) {
-        sslHandler = new SslHandler(helper.createEngine(vertx, remoteAddress, serverName, false));
-        ((SslHandler) sslHandler).setHandshakeTimeout(helper.getSslHandshakeTimeout(), helper.getSslHandshakeTimeoutUnit());
+        sslHandler = helper.createSslHandler(vertx, remoteAddress, serverName, false);
       } else {
         if (helper.isSNI()) {
           sslHandler = new SniHandler(helper.serverNameMapper(vertx));
         } else {
-          sslHandler = new SslHandler(helper.createEngine(vertx));
-          ((SslHandler) sslHandler).setHandshakeTimeout(helper.getSslHandshakeTimeout(), helper.getSslHandshakeTimeoutUnit());
+          sslHandler = helper.createSslHandler(vertx, null, null);
         }
       }
       chctx.pipeline().addFirst("ssl", sslHandler);
