+++ src/main/java/io/vertx/core/spi/tls/SslContextFactory.java
@@ -1,0 +1,14 @@
+package io.vertx.core.spi.tls;
+
+import io.netty.handler.ssl.SslContext;
+import io.vertx.core.impl.VertxInternal;
+
+public interface SslContextFactory {
+
+  SslContext getContext(VertxInternal vertx,
+                        String serverName,
+                        boolean useAlpn,
+                        boolean client,
+                        boolean trustAll);
+
+}
